Fix desktop SOS tests
